src/requirements.txt
azure-functions
azure.identity
openai
pydantic-settings